+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - added: Buy conversion tracking for Moonpay
 - added: Error tracking for failure to report conversions to referral server
 - changed: Remove whitespaces from custom token contract address input
+- changed: Use unique ENV configs for thorchain and thorchainda swap plugins
 - fixed: Correctly tag `tokenApproval` `actionType` in  `getTxActionDisplayInfo`
 - fixed: AddressTile2 touchable area states
 - fixed: Cases where it was possible to create duplicate custom tokens